Migrated 'allowed-sharing-modes'.
